refact: Class Pinyin for extend and override segment method
